Change max zenith to 89.999
Updated gensky to keep the sun position 0.001 degrees from the zenith instead of 3.0 degrees from the zenith. This 3 degree error causes inaccuracies in locations between the tropics where the sun can be truly directly overhead. A difference of 0.001 degrees is approximately the same as a 1km offset in position at the equator, a 3 degree error is the equivalent of having the model be more than 300km than where it should be.